don't hack scope
